// src/spa/src/app/basket/models/basket-view-model.ts
import { BasketForCheckoutDto } from 'src/dto/models';
import { LibraryAssetForBasketViewModel } from './library-asset-for-basket-view-model';

export interface BasketViewModel extends BasketForCheckoutDto {
  cardNumber: string;
  photoUrl: string;
  assets: LibraryAssetForBasketViewModel[];
  active: boolean;
  checkoutComplete: boolean;
}